requirements.txt
python-telegram-bot==21.5
requests==2.31.0
schedule==1.2.1
psycopg2-binary==2.9.9
